requirements.txt
Django==5.0.6
python-dotenv==1.0.1
django-ipware==7.0.1
whitenoise==6.7.0
gunicorn
